src/requirements.txt
streamlit==1.46.1
fastapi==0.115.2
uvicorn==0.32.0
groq==0.4.1
requests==2.31.0
pymongo==4.8.0
python-dateutil==2.8.2
plotly==5.24.1
httpx==0.27.2
python-dotenv==1.1.1
google-generativeai==0.8.3
Pillow==10.4.0
langgraph==0.2.39
langchain-core==0.3.12
tenacity==8.2.3
motor==3.5.1
python-multipart
logging
